embedding-training/requirements.txt
fastapi
uvicorn
psycopg2-binary
transformers
torch
optimum[onnxruntime]
pydantic
pydantic-settings
jinja2>=2.8